MedicalChatbot/requirements.txt
python=3.11
google-generativeai
streamlit